prevented teardown failure
git-svn-id: svn://pc-lab14/SVN/cloudify/trunk/quality/frameworks/SGTest@137881 eb64e737-3616-4df0-8941-5ee2ae88103d

+++ src/main/java/test/cli/cloudify/cloud/byon/DisconnectAndReconnectAgentMachineTest.java
@@ -62,7 +62,7 @@ public class DisconnectAndReconnectAgentMachineTest extends AbstractByonCloudTes
 			WANemUtils.removeRoutingTableEntries(ips);
 		}
 		
-		if(machineToDisconnect != null){
+		if(machineToDisconnect != null && !ServiceUtils.isPortFree(machineToDisconnect.getHostAddress(), SERVICE_PORT)){
 			SSHUtils.runCommand(machineToDisconnect.getHostAddress(), OPERATION_TIMEOUT/3, "killall -9 " + SERVICE_NAME, USER, PASSWORD);
 		}
